Don't create two connections per test
A hidden connection is created by new Socket()
It automatically connects.  So the open() call creates a second
connection.  Specifying {manual:true} corrects that.

Optionally, {manual:true} can be left out and the call to open() can be
removed.  That way we'd rely on the implicit (hidden) connection from
new Socket().  I prefer to keep the connection open explicit so that one
doesn't have to wonder whether the callbacks have time to be registered
before the implicit connection is opened.  Looking at the source, they
do have time via a setTimeout for the implicit open() call but one
wouldn't know that by looking at the client code.